+++ examples/instance_template/additional_disks/main.tf
@@ -30,18 +30,24 @@ module "instance_template" {
 
   additional_disks = [
     {
+      disk_name    = "disk-0"
+      device_name  = "disk-0"
       disk_size_gb = 10
       disk_type    = "pd-standard"
       auto_delete  = "true"
       boot         = "false"
     },
     {
+      disk_name    = "disk-1"
+      device_name  = "disk-1"
       disk_size_gb = 10
       disk_type    = "pd-standard"
       auto_delete  = "true"
       boot         = "false"
     },
     {
+      disk_name    = "disk-2"
+      device_name  = "disk-2"
       disk_size_gb = 10
       disk_type    = "pd-standard"
       auto_delete  = "true"